comment out H1 and template list, rename splash
