+++ src/components/Splash/index.js
